% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{allParams}
\alias{allParams}
\title{All Parameters}
\arguments{
\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
A character vector of parameter key names
}
\description{
ParamSetI object method to list all parameter keys
}
\details{
\preformatted{
## Method Usage:
myObject$allParams( help=TRUE )

myObject$allParams()
}

Lists all "known" parameter key names. This includes parameters
that have been set, as well as those that are unset but have been
described.
}
\examples{

## This demo defines a toy object inheriting ParamSetI:
demo("exampleParamSetObject", package="ParamSetI", ask=FALSE)
pse <- ParamSetExample( params=list(inc=10L) )

pse$allParams()
pse$param("puppy", "fuzzy") # Add a new parameter
pse$allParams()

}
\seealso{
\link{hasParam}, \link{showParameters}, \link{param}
}
