% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{defineParameters}
\alias{defineParameters}
\title{Define Parameters Set Param List}
\arguments{
\item{x}{Required, the block of text to parse (character)}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\description{
ParamSetI object method to bulk set definitions (descriptions) for parameters
}
\details{
\preformatted{
## Method Usage:
myObject$defineParameters( help=TRUE )

myObject$defineParameters( x )
}

A mechanism to bulk set parameters using a block of text. The text
block will be split into lines and parsed with regular expressions
to extract names, class restrictions and definitions. The method
does not define values.
}
\examples{

## This demo defines a toy object inheriting ParamSetI:
demo("exampleParamSetObject", package="ParamSetI", ask=FALSE)
pse <- ParamSetExample( params=list(inc=10L) )

# Define four new parameters, three of which have class
# restrictions, one does not:
pse$defineParameters("
   color [character] The color of the object, a name or hex value
  weight [numeric]   The mass of the object, in kilograms
 inStock [logical]   Flag to indicate if inventory is available
    misc Random information, see Jacob in logistics for more info
")

# Set a few of these
pse$param("weight", 74.3)
pse$param("instock", "yes") # Oops, violates class restriction
pse$param("misc", "[/][Stack on lower shelf/No hooks]")

# Show all the definitions
pse$showParameters( na.rm=FALSE )

}
\seealso{
\link{param}, \link{setParamList}, \link{showParameters}
}
