% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{hasParam}
\alias{hasParam}
\title{Has Parameter}
\arguments{
\item{key}{Default NULL. One or more parameter names}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
A logical vector with TRUE indicating presence of the
    corresponding key in the parameter set
}
\description{
ParamSetI object method to test if an object has a parameter or not
}
\details{
\preformatted{
## Method Usage:
myObject$hasParam( help=TRUE )

myObject$hasParam( key=NULL )
}

Checks if the provided key(s) are defined in the parameter
set. Case/capitalization is ignored.
}
\examples{

## This demo defines a toy object inheriting ParamSetI:
demo("exampleParamSetObject", package="ParamSetI", ask=FALSE)
pse <- ParamSetExample( params=list(inc=10L) )

pse$showParameters()
pse$hasParam("inc")
pse$hasParam("maximum operating depth")

}
\seealso{
\link{allParams}
}
