% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{paramDefinition}
\alias{paramDefinition}
\title{Parameter Definition}
\arguments{
\item{key}{Required, one or more parameter names}

\item{val}{Default NULL. Optional character vector of new parameter
definitions}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
A named character vector of definitions, with names
    corresponding to keys
}
\description{
ParamSetI object method to get/set the definition (description) of a parameter
}
\details{
\preformatted{
## Method Usage:
myObject$paramDefinition( help=TRUE )

myObject$paramDefinition( key, val=NULL )
}

Parameters can have optional descriptions to help inform users of
their usage. This function returns these definitions, or allows
them to be set. Deffinitions will be displayed in some output
functions, such as \link{showParameters}.

For setting large numbers of parameters, also consider
\link{defineParameters}, which can parse descriptions from a block of
text.
}
\examples{

## This demo defines a toy object inheriting ParamSetI:
demo("exampleParamSetObject", package="ParamSetI", ask=FALSE)
pse <- ParamSetExample( params=list(inc=10L) )

# Get the definition for a parameter
pse$paramDefinition("inc")
# Set the definitions for some new parameters
pse$paramDefinition(c("color", "velocity"),
                    c("The color of the object", "The speed of the object"))
# Show all descriptions
pse$paramDefinition( pse$allParams() )

}
\seealso{
\link{defineParameters}, \link{allParams}, \link{showParameters}
}
