% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{paramName}
\alias{paramName}
\title{Parameter Name}
\arguments{
\item{key}{Required, one or more parameter names}

\item{val}{Default NULL. Optional character vector of new parameter
names. \code{val} can only differ from \code{key} in the
capitalization, otherwise an error will be thrown.}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
A named character vector of class restrictions, with names
    corresponding to keys
}
\description{
ParamSetI object method to get/set the pretty-print name of a parameter
}
\details{
\preformatted{
## Method Usage:
myObject$paramName( help=TRUE )

myObject$paramName( key, val=NULL )
}

A mostly-aesthetic function to configure the case/capitalization of
parameter names. Internally the package ignores case assigned to
parameters. However, for display purposes the capitalization can be
specifically set. This method allows a particular capitalization to
be defined.
}
\examples{

## This demo defines a toy object inheriting ParamSetI:
demo("exampleParamSetObject", package="ParamSetI", ask=FALSE)
pse <- ParamSetExample( params=list(inc=10L) )

# Get the pretty print name for a parameter
pse$paramClass("inc")

# That's boring. SHOUT IT
pse$paramClass("inc", "INC")

# Mostly handy for CamelCase. Note that key can be any case
pse$param("ADARKANDSTORMYNIGHT", "aDarkAndStormyNight")
}
\seealso{
\link{defineParameters}, \link{showParameters}
}
