% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamSetI.R
\docType{class}
\name{ParamSetI-class}
\alias{ParamSetI-class}
\alias{ParamSetI}
\title{Paramter Set Interface}
\description{
Inherited class holding parameter manipulation functions
}
\section{Fields}{

\describe{
\item{\code{paramSet}}{List holding key-value pairs used by an object}

\item{\code{paramDef}}{List holding parameter definitions and class
restrictions}

\item{\code{varName}}{Single string holding extracted variable names. Used
for reporting R code that utilizes actual variable names being
used.}
}}

\section{Methods}{

\describe{
\item{\code{allParams(help = FALSE)}}{Return all parameter keys, including those that are set or just described}

\item{\code{defineParameters(x, help = FALSE)}}{Bulk set parameter definitions with a block of text}

\item{\code{hasParam(key = NULL, help = FALSE)}}{Test if one or more parameter keys exist in the object}

\item{\code{initialize(useColor = NULL, verbose = NULL, log = NULL, ...)}}{\preformatted{
Create a new object using EventLogger():
        log - Optional EventLogger object. This is used if EventLogger is an
              inherited ('contains') class in another RefClass object, and you
              wish that object to share the event log from a previously created
              object.
   useColor - Defaults to TRUE, toggles if crayon::-based coloring is applied
              to printed messages. If 'log' is provided, changes will affect
              that shared object
    verbose - Defaults to TRUE, toggles if log events are also printed to the
              terminal. If 'log' is provided, changes will affect that shared
              object
}}

\item{\code{param(key = NA, val = NA, append = FALSE, default = NA,
  clobber = TRUE, check.class = NULL, is.scalar = NULL, coerce = TRUE,
  help = FALSE)}}{Get or set a parameter}

\item{\code{paramClass(key, val = NULL, help = FALSE)}}{Get or set class (storage mode) restrictions of parameters}

\item{\code{paramDefinition(key, val = NULL, help = FALSE)}}{Return the definitions, if any, for one or more parameter keys}

\item{\code{paramName(key, val = NULL, help = FALSE)}}{Set the capitalization to use for a parameter name}

\item{\code{setParamList(params = NULL, help = FALSE, ...)}}{Set multiple parameters using a list}

\item{\code{showParameters(na.rm = TRUE, help = FALSE)}}{Pretty-print parameter names, values and definitions}
}}

\examples{

## Demo with inheritance of the class:
demo("objectInheritance", package="ParamSetI", ask=FALSE)

}
