% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamSetI.R
\name{selfSplittingString}
\alias{selfSplittingString}
\title{Self-Splitting String}
\usage{
selfSplittingString(x)
}
\arguments{
\item{x}{Required, the string vector to proces}
}
\value{
NULL if passed NULL, otherwise a character vector
}
\description{
Parses a string that defines a split token to turn it into a vector
}
\details{
Helper function for ParamSetI, takes a character vector as
input. If the input is length 1 and is of format:

   [<TOKEN>][<TEXT>]

... then <TEXT> will be split into a vector by <TOKEN>. For example:

   [/][x/  y  /z]

... becomes c("x", "  y  ", "z").

In all other cases, only a single value will be returned,
corresponding to the first non-NA value in x.
}
\examples{

selfSplittingString(" [,][a,b, c ]  ")
# "a"   "b"   " c "

selfSplittingString("[ and ][lions and tigers and bears oh my]")
# "lions" "tigers" "bears oh my"

# A single return value (vector length 1) is enforced:
selfSplittingString(c("a","b","c"))
# "a"

}
