% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{setParamList}
\alias{setParamList}
\title{Set Param List}
\arguments{
\item{params}{Default NULL, which will be a no-op and return
NA. Otherwise, a list holding the parameters, with the names
representing key names and the contents the values to assign to
each key}

\item{...}{Any other parameters will be passed to param() for each
key/value pair}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\description{
ParamSetI object method to set parameters in bulk using a list object
}
\details{
\preformatted{
## Method Usage:
myObject$setParamList( help=TRUE )

myObject$setParamList( params=NULL, ... )
}

A mechanism to bulk set parameters using a list. The list names
will be taken as keys, and their contents as values.
}
\examples{

## This demo defines a toy object inheriting ParamSetI:
demo("exampleParamSetObject", package="ParamSetI", ask=FALSE)
pse <- ParamSetExample( params=list(inc=10L) )

# What is already set:
pse$showParameters()

# Reset some existing parameters:
pse$setParamList( list( inc=7L, dec=2L ) )
# Set some new parameters:
pse$setParamList( list(color="green", speed="88mph") )
# Check current values
pse$showParameters()

# Can unset clobber to set defaults without overwriting current values
pse$setParamList( list( inc=999L, dec=888L ), clobber=FALSE )
pse$showParameters()

}
\seealso{
\link{defineParameters}, \link{param},
    \link{showParameters}
}
