% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{showParameters}
\alias{showParameters}
\title{Show Parameters}
\arguments{
\item{na.rm}{Default TRUE, which will exclude parameters that are
not "defined" (see \link[CatMisc]{is.def} for values that are
considered 'undefined'). Setting to FALSE shows all parameters
available from \link{allParams}.}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\description{
ParamSetI object method to pretty-print parameters and their values
}
\details{
\preformatted{
## Method Usage:
myObject$showParameters( help=TRUE )

myObject$showParameters( na.rm=TRUE )
}

Show parameters in a format designed for human
consumption. Parameters will be shown wrapped in a \code{$param()}
call to illustrate how to access and change the parameter. If a
definition has been set, it will be shown as a comment line under
the parameter value.
}
\examples{

## This demo defines a toy object inheriting ParamSetI:
demo("exampleParamSetObject", package="ParamSetI", ask=FALSE)
pse <- ParamSetExample( params=list(inc=10L) )

# Show all parameters with defined values
pse$showParameters(  )

# Show all parameters, including one that has an undefined value
# but an assigned deffinition (description)
pse$showParameters( na.rm=FALSE )

}
\seealso{
\link{param}
}
